app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cohere==5.11.2
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0
fastavro==1.9.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.6
langchain-cohere==0.3.1
langchain-community==0.3.4
langchain-core==0.3.15
langchain-experimental==0.3.2
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.139
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.53.0
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-alephalpha==0.33.12
opentelemetry-instrumentation-anthropic==0.33.12
opentelemetry-instrumentation-bedrock==0.33.12
opentelemetry-instrumentation-chromadb==0.33.12
opentelemetry-instrumentation-cohere==0.33.12
opentelemetry-instrumentation-google-generativeai==0.33.12
opentelemetry-instrumentation-groq==0.33.12
opentelemetry-instrumentation-haystack==0.33.12
opentelemetry-instrumentation-lancedb==0.33.12
opentelemetry-instrumentation-langchain==0.33.12
opentelemetry-instrumentation-llamaindex==0.33.12
opentelemetry-instrumentation-logging==0.49b1
opentelemetry-instrumentation-marqo==0.33.12
opentelemetry-instrumentation-milvus==0.33.12
opentelemetry-instrumentation-mistralai==0.33.12
opentelemetry-instrumentation-ollama==0.33.12
opentelemetry-instrumentation-openai==0.33.12
opentelemetry-instrumentation-pinecone==0.33.12
opentelemetry-instrumentation-qdrant==0.33.12
opentelemetry-instrumentation-replicate==0.33.12
opentelemetry-instrumentation-requests==0.49b1
opentelemetry-instrumentation-sagemaker==0.33.12
opentelemetry-instrumentation-sqlalchemy==0.49b1
opentelemetry-instrumentation-threading==0.49b1
opentelemetry-instrumentation-together==0.33.12
opentelemetry-instrumentation-transformers==0.33.12
opentelemetry-instrumentation-urllib3==0.49b1
opentelemetry-instrumentation-vertexai==0.33.12
opentelemetry-instrumentation-watsonx==0.33.12
opentelemetry-instrumentation-weaviate==0.33.12
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.49b1
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.40.1
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
traceloop-sdk==0.33.12
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
